+++ resources/js/components/screens/transactions/index.ts
@@ -1,0 +1,2 @@
+export { default as Filters } from './Filters.vue';
+export { default as List } from './List.vue';
