+++ resources/js/types/utils.d.ts
@@ -1,0 +1,9 @@
+export type TTransactionFilters = {
+    search: string;
+    kind: string;
+    category: string;
+    dateFrom: string;
+    dateTo: string;
+    minAmount: string;
+    maxAmount: string;
+};
